refactor for better code structure
