+++ src/main/java/com/example/BookingSystem/Controllers/BookingController.java
@@ -1,5 +1,7 @@
-package com.example.BookingSystem.Booking;
+package com.example.BookingSystem.Controllers;
 
+import com.example.BookingSystem.Models.Booking;
+import com.example.BookingSystem.Services.BookingService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
