+++ src/main/java/com/example/BookingSystem/Services/BookingService.java
@@ -1,5 +1,7 @@
-package com.example.BookingSystem.Booking;
+package com.example.BookingSystem.Services;
 
+import com.example.BookingSystem.Models.Booking;
+import com.example.BookingSystem.Repositories.BookingRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
